skew is now a double instead of an int, and first pass at implementing neume rendering with staff skew

+++ include/vrv/facsimileinterface.h
@@ -38,7 +38,7 @@ public:
     int GetWidth() const;
     int GetHeight() const;
 
-    int GetDrawingSkew() const;
+    double GetDrawingSkew() const;
 
     /** Check if the object has a facsimile */
     bool HasFacsimile() { return this->HasFacs(); }
